[refactor] Extract `wp_kubernetes` module
